Moved user model to /models/user-model.js
