+++ server/userModel/userData.js
@@ -1,4 +1,4 @@
-const User = require("./user");
+const User = require("../models/user-model");
 const bcrypt = require("bcrypt");
 const saltRounds = 10;
 const jwt = require("jsonwebtoken");
